fix: configured validation for CentralApi
Moved ValidationExceptionMiddleware in Common and used it for CentralApi
