+++ src/CentralApi/Program.cs
@@ -1,6 +1,7 @@
 using Middleware.CentralApi.ExtensionMethods;
 using Middleware.CentralApi.Services;
 using Middleware.Common.ExtensionMethods;
+using Middleware.Common.Validation;
 using Middleware.DataAccess.ExtensionMethods;
 using Middleware.DataAccess.HostedServices;
 using Middleware.DataAccess.Repositories;
@@ -15,11 +16,14 @@ builder.ConfigureLogger();
 builder.RegisterRedis();
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+
+builder.Services.AddFluentValidation(typeof(Program));
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 builder.Services.AddHostedService<IndexCreationService>();
 builder.Services.AddScoped<ICloudRepository, RedisCloudRepository>();
 builder.Services.AddScoped<IEdgeRepository, RedisEdgeRepository>();
+builder.Services.AddScoped<ILocationRepository, RedisLocationRepository>();
 builder.Services.AddScoped<ILocationService, LocationService>();
 
 builder.Services.RegisterCentralApiQuartzJobs();
@@ -36,6 +40,8 @@ if (app.Environment.IsDevelopment())
 //app.UseHttpsRedirection();
 
 //app.UseAuthorization();
+
+app.UseMiddleware<ValidationExceptionMiddleware>();
 
 app.MapControllers();
