+++ src/RedisInterface/Program.cs
@@ -2,6 +2,7 @@ using System.Reflection;
 using Middleware.CentralApi.Sdk;
 using Middleware.Common.Config;
 using Middleware.Common.ExtensionMethods;
+using Middleware.Common.Validation;
 using Middleware.DataAccess.ExtensionMethods;
 using Middleware.DataAccess.HostedServices;
 using Middleware.RedisInterface;
